# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Manga service api",
        "title": "Hanan Test",
        "contact": {
            "name": "API manga",
            "email": "hanan@asyrawih.id"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/manga": {
            "get": {
                "description": "get all manga",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manga"
                ],
                "summary": "List manga",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Manga"
                            }
                        }
                    }
                }
            }
        },
        "/manga/{id}": {
            "get": {
                "description": "get all manga",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manga"
                ],
                "summary": "List manga",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Manga Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Manga"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Manga based on id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manga"
                ],
                "summary": "Remove manga",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Manga Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Manga": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "release_date": {
                    "type": "string"
                },
                "sinopsis": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "total_chapter": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        }
    }
}